My Biggest Push.
Set up precurrsors to automatic Excel sheet making

EmailDump.py, FileDump.py, URLDump.py
collects emails urls and files found in triage in globals for later use

the spec.pys are going to be used and called in Excel sheet exporting. used to read from the globals

GlobalChecker is going to be a checker to see if user has enough knowledge to make a MISP report.

Fixed Formatting to look better for usage. main enter point. and Error.

added a yes or no to make sure you want to exit

thats ALL! on to the next